requirements.txt
pandas==2.2.2
openpyxl==3.1.2
requests==2.32.3
jsonschema==4.22.0
python-dateutil==2.9.0.post0
openai>=1.0.0
feedparser==6.0.11
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.3.0
lxml_html_clean==0.1.1
beautifulsoup4
feedparser
jsonschema
lxml
lxml_html_clean
openpyxl
pandas
python-dateutil
readability-lxml
requests
streamlit
apscheduler
markdown
weasyprint
markdown2
